reorganized base classes in own module
