+++ plugins/roles.py
@@ -5,7 +5,7 @@ import discord
 import emoji
 from discord.ext import commands
 
-from Geckarbot import BasePlugin
+from base import BasePlugin
 from conf import Config
 from botutils import utils, permChecks
 from subsystems import reactions
